Moved request and response classes to separate folders
