+++ src/Message/Request/CompletePurchaseRequest.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+
+namespace Omnipay\Utrust\Message\Request;
+
+class CompletePurchaseRequest extends AbstractRequest
+{
+    public function getData()
+    {
+        // TODO: Implement getData() method.
+    }
+
+    protected function getPath(): ?string
+    {
+        // TODO: Implement getPath() method.
+    }
+
+    protected function getResponseClass(): string
+    {
+        // TODO: Implement getResponseClass() method.
+    }
+}
