+++ src/Message/Response/AbstractResponse.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Omnipay\Utrust\Message\Response;
+
+abstract class AbstractResponse extends \Omnipay\Common\Message\AbstractResponse
+{
+
+}
