+++ src/Message/Response/PurchaseResponse.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Omnipay\Utrust\Message\Response;
+
+class PurchaseResponse extends AbstractResponse
+{
+    public function isSuccessful()
+    {
+        // TODO: Implement isSuccessful() method.
+    }
+}
